Mileage Tracker Models (12/4/19)
These two models were created for the two tickets I am assigned as of
this week. Names (Miles | Commute)

Miles:
This model has a user which is a `user` class that is imported from
'django.contrib.auth.models'.
Also in this model is `miles` which is a float field
(number with a decimal)

Commute:
This model has a user which is a `user` class that is imported from
'django.contrib.auth.models'.
Also in this model is `commute` which is a text field (string).