mileage_tracker/views.py & /forms.py
views.py:
The previous function 'set_commute' was changed to a
CreateView class. NEW NAME: DistanceToWorkCreateView
ARGUMENTS: LoginRequiredMixin, CreateView
This class does the same thing as the previous function, but now
allows for the user to be able to change their commute distance.

forms.py:
This file was deleted because using a CreateView creates the form it
self, rendering the form useless.